examples/google-adk/pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Example AI agent for AWS Bedrock AgentCore"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.15.0",
    "litellm>=1.77.0",
    "pyyaml>=6.0",
    "boto3>=1.40.0",
    "bedrock-agentcore>=0.1.7",
    # OpenTelemetry for AWS X-Ray tracing (compatible with google-adk's OTEL deps)
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
]
